+++ examples/sklearn/iris_skl.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-fast-deploy sklearn \
+quick-deploy sklearn \
     -n iris_cls \
     -m iris_cls.bin \
     -o ./models \