+++ examples/timms/timm_resnet18.sh
@@ -2,7 +2,7 @@
 
 python -c "import timm; import torch; torch.save(timm.create_model('resnet18', pretrained=True), 'resnet18.pt')"
 
-fast-deploy torch \
+quick-deploy torch \
     -n resnet18 \
     -m resnet18.pt \
     -o ./models/resnet18 \